@@ -4,16 +4,18 @@
   "private": true,
   "proxy": "http://localhost:3001",
   "dependencies": {
+    "@apollo/client": "^3.5.10",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "bootstrap": "^4.4.1",
-    "jwt-decode": "^2.2.0",
-    "react": "^16.13.1",
+    "graphql": "^16.3.0",
+    "jwt-decode": "^3.1.2",
+    "react": "^18.0.0",
     "react-bootstrap": "^1.0.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1"
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
