@@ -8,49 +8,108 @@
   ],
   "files": [
     "lib",
-    "index.js",
-    "browser_build",
+    "src",
+    "dist",
+    "bson.d.ts",
+    "etc/prepare.js",
     "bower.json"
   ],
-  "version": "1.1.4",
-  "author": "Christian Amor Kvalheim <christkv@gmail.com>",
+  "types": "bson.d.ts",
+  "version": "4.6.2",
+  "author": {
+    "name": "The MongoDB NodeJS Team",
+    "email": "dbx-node@mongodb.com"
+  },
+  "license": "Apache-2.0",
   "contributors": [],
   "repository": "mongodb/js-bson",
   "bugs": {
-    "mail": "node-mongodb-native@googlegroups.com",
-    "url": "https://github.com/mongodb/js-bson/issues"
+    "url": "https://jira.mongodb.org/projects/NODE/issues/"
   },
   "devDependencies": {
-    "benchmark": "1.0.0",
-    "colors": "1.1.0",
-    "nodeunit": "0.9.0",
-    "babel-core": "^6.14.0",
-    "babel-loader": "^6.2.5",
-    "babel-polyfill": "^6.13.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.14.0",
-    "conventional-changelog-cli": "^1.3.5",
-    "webpack": "^1.13.2",
-    "webpack-polyfills-plugin": "0.0.9"
+    "@babel/plugin-external-helpers": "^7.10.4",
+    "@babel/preset-env": "^7.11.0",
+    "@istanbuljs/nyc-config-typescript": "^1.0.1",
+    "@microsoft/api-extractor": "^7.11.2",
+    "@rollup/plugin-babel": "^5.2.0",
+    "@rollup/plugin-commonjs": "^15.0.0",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^9.0.0",
+    "@rollup/plugin-typescript": "^6.0.0",
+    "@typescript-eslint/eslint-plugin": "^3.10.1",
+    "@typescript-eslint/parser": "^3.10.1",
+    "array-includes": "^3.1.3",
+    "benchmark": "^2.1.4",
+    "chai": "^4.2.0",
+    "downlevel-dts": "^0.7.0",
+    "eslint": "^7.7.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-tsdoc": "^0.2.6",
+    "karma": "^6.3.4",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-mocha": "^2.0.1",
+    "karma-mocha-reporter": "^2.2.5",
+    "karma-rollup-preprocessor": "^7.0.5",
+    "mocha": "5.2.0",
+    "node-fetch": "^2.6.1",
+    "nyc": "^15.1.0",
+    "object.entries": "^1.1.4",
+    "prettier": "^2.1.1",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.26.5",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-globals": "^1.4.0",
+    "rollup-plugin-node-polyfills": "^0.2.1",
+    "rollup-plugin-polyfill-node": "^0.7.0",
+    "standard-version": "^9.3.0",
+    "ts-node": "^9.0.0",
+    "tsd": "^0.17.0",
+    "typedoc": "^0.21.2",
+    "typescript": "^4.0.2",
+    "typescript-cached-transpile": "0.0.6",
+    "uuid": "^8.3.2"
+  },
+  "tsd": {
+    "directory": "test/types",
+    "compilerOptions": {
+      "strict": true,
+      "target": "esnext",
+      "module": "commonjs",
+      "moduleResolution": "node"
+    }
   },
   "config": {
     "native": false
   },
-  "main": "./index",
-  "directories": {
-    "lib": "./lib/bson"
+  "main": "lib/bson.js",
+  "module": "dist/bson.esm.js",
+  "browser": {
+    "./lib/bson.js": "./dist/bson.browser.umd.js",
+    "./dist/bson.esm.js": "./dist/bson.browser.esm.js"
   },
   "engines": {
-    "node": ">=0.6.19"
+    "node": ">=6.9.0"
   },
   "scripts": {
-    "test": "nodeunit ./test/node",
-    "build": "webpack --config ./webpack.dist.config.js",
-    "changelog": "conventional-changelog -p angular -i HISTORY.md -s",
-    "lint": "eslint lib test",
-    "format": "prettier --print-width 100 --tab-width 2 --single-quote --write 'test/**/*.js' 'lib/**/*.js'"
+    "docs": "typedoc",
+    "test": "npm run build && npm run test-node && npm run test-browser",
+    "test-node": "mocha test/node test/*_tests.js",
+    "test-tsd": "npm run build:dts && tsd",
+    "test-browser": "node --max-old-space-size=4096 ./node_modules/.bin/karma start karma.conf.js",
+    "build:ts": "tsc",
+    "build:dts": "npm run build:ts && api-extractor run --typescript-compiler-folder node_modules/typescript --local && rimraf 'lib/**/*.d.ts*' && downlevel-dts bson.d.ts bson.d.ts",
+    "build:bundle": "rollup -c rollup.config.js",
+    "build": "npm run build:dts && npm run build:bundle",
+    "lint": "eslint -v && eslint --ext '.js,.ts' --max-warnings=0 src test && tsc -v && tsc --noEmit && npm run test-tsd",
+    "format": "eslint --ext '.js,.ts' src test --fix",
+    "coverage": "nyc npm run test-node",
+    "coverage:html": "npm run coverage && open ./coverage/index.html",
+    "prepare": "node etc/prepare.js",
+    "release": "standard-version -i HISTORY.md"
   },
-  "browser": "lib/bson/bson.js",
-  "license": "Apache-2.0"
+  "dependencies": {
+    "buffer": "^5.6.0"
+  }
 }
