@@ -1,11 +1,22 @@
 {
   "name": "graphql",
-  "version": "15.5.0",
+  "version": "16.3.0",
   "description": "A Query Language and Runtime which can target any service.",
   "license": "MIT",
   "main": "index",
   "module": "index.mjs",
-  "types": "index.d.ts",
+  "typesVersions": {
+    ">=4.1.0": {
+      "*": [
+        "*"
+      ]
+    },
+    "*": {
+      "*": [
+        "NotSupportedTSVersion.d.ts"
+      ]
+    }
+  },
   "sideEffects": false,
   "homepage": "https://github.com/graphql/graphql-js",
   "bugs": {
@@ -20,6 +31,9 @@
     "graphql-js"
   ],
   "engines": {
-    "node": ">= 10.x"
+    "node": "^12.22.0 || ^14.16.0 || >=16.0.0"
+  },
+  "publishConfig": {
+    "tag": "latest"
   }
 }
