@@ -16,66 +27,108 @@
     "driver",
     "official"
   ],
-  "peerOptionalDependencies": {
-    "kerberos": "^1.1.0",
-    "mongodb-client-encryption": "^1.0.0",
-    "mongodb-extjson": "^2.1.2",
-    "snappy": "^6.1.1",
-    "bson-ext": "^2.0.0"
+  "author": {
+    "name": "The MongoDB NodeJS Team",
+    "email": "dbx-node@mongodb.com"
   },
   "dependencies": {
-    "bl": "^2.2.0",
-    "bson": "^1.1.4",
-    "denque": "^1.4.1",
-    "require_optional": "^1.0.1",
-    "safe-buffer": "^5.1.2"
+    "bson": "^4.6.1",
+    "denque": "^2.0.1",
+    "mongodb-connection-string-url": "^2.4.1",
+    "socks": "^2.6.1"
   },
   "devDependencies": {
-    "chai": "^4.1.1",
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
+    "@microsoft/api-extractor": "^7.19.4",
+    "@microsoft/tsdoc-config": "^0.15.2",
+    "@types/chai": "^4.3.0",
+    "@types/chai-subset": "^1.3.3",
+    "@types/kerberos": "^1.1.1",
+    "@types/mocha": "^9.0.0",
+    "@types/node": "^17.0.10",
+    "@types/saslprep": "^1.0.1",
+    "@types/semver": "^7.3.9",
+    "@types/sinon": "^10.0.6",
+    "@types/whatwg-url": "^8.2.1",
+    "@typescript-eslint/eslint-plugin": "^5.10.0",
+    "@typescript-eslint/parser": "^5.10.0",
+    "bluebird": "^3.7.2",
+    "chai": "^4.3.4",
     "chai-subset": "^1.6.0",
-    "chalk": "^2.4.2",
-    "co": "4.6.0",
-    "coveralls": "^2.11.6",
-    "eslint": "^4.5.0",
-    "eslint-plugin-prettier": "^2.2.0",
-    "istanbul": "^0.4.5",
-    "jsdoc": "3.5.5",
+    "chalk": "^4.1.2",
+    "downlevel-dts": "^0.8.0",
+    "eslint": "^8.7.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "eslint-plugin-tsdoc": "^0.2.14",
+    "js-yaml": "^4.1.0",
     "lodash.camelcase": "^4.3.0",
-    "mocha": "5.2.0",
-    "mocha-sinon": "^2.1.0",
-    "mongodb-extjson": "^2.1.1",
-    "mongodb-mock-server": "^1.0.1",
-    "prettier": "^1.19.1",
-    "semver": "^5.5.0",
-    "sinon": "^4.3.0",
-    "sinon-chai": "^3.2.0",
-    "snappy": "^6.1.2",
-    "standard-version": "^4.4.0",
-    "util.promisify": "^1.0.1",
-    "worker-farm": "^1.5.0",
-    "wtfnode": "^0.8.0",
-    "yargs": "^14.2.0"
+    "mocha": "^9.1.4",
+    "mocha-sinon": "^2.1.2",
+    "nyc": "^15.1.0",
+    "prettier": "^2.5.1",
+    "rimraf": "^3.0.2",
+    "semver": "^7.3.5",
+    "sinon": "^12.0.1",
+    "sinon-chai": "^3.7.0",
+    "source-map-support": "^0.5.21",
+    "standard-version": "^9.3.2",
+    "ts-node": "^10.4.0",
+    "tsd": "^0.19.1",
+    "typescript": "^4.5.4",
+    "typescript-cached-transpile": "^0.0.6",
+    "xml2js": "^0.4.23",
+    "yargs": "^17.3.1"
   },
   "license": "Apache-2.0",
   "engines": {
-    "node": ">=4"
+    "node": ">=12.9.0"
   },
   "bugs": {
-    "url": "https://github.com/mongodb/node-mongodb-native/issues"
-  },
-  "scripts": {
-    "atlas": "node ./test/tools/atlas_connectivity_tests.js",
-    "test": "npm run lint && mocha --recursive test/functional test/unit test/core",
-    "test-nolint": "mocha --recursive test/functional test/unit test/core",
-    "coverage": "istanbul cover mongodb-test-runner -- -t 60000 test/core test/unit test/functional",
-    "lint": "eslint lib test",
-    "format": "prettier --print-width 100 --tab-width 2 --single-quote --write 'test/**/*.js' 'lib/**/*.js'",
-    "bench": "node test/benchmarks/driverBench/",
-    "generate-evergreen": "node .evergreen/generate_evergreen_tasks.js",
-    "release": "standard-version -i HISTORY.md"
+    "url": "https://jira.mongodb.org/projects/NODE/issues/"
   },
   "homepage": "https://github.com/mongodb/node-mongodb-native",
   "optionalDependencies": {
-    "saslprep": "^1.0.0"
+    "saslprep": "^1.0.3"
+  },
+  "scripts": {
+    "build:evergreen": "node .evergreen/generate_evergreen_tasks.js",
+    "build:ts": "rimraf lib && ./node_modules/typescript/bin/tsc",
+    "build:dts": "npm run build:ts && api-extractor run && rimraf 'lib/**/*.d.ts*' && downlevel-dts mongodb.d.ts mongodb.ts34.d.ts",
+    "build:docs": "typedoc",
+    "check:bench": "node test/benchmarks/driverBench",
+    "check:coverage": "nyc npm run test:all",
+    "check:lint": "npm run build:dts && npm run check:dts && npm run check:eslint && npm run check:tsd",
+    "check:eslint": "eslint -v && eslint --max-warnings=0 --ext '.js,.ts' src test",
+    "check:tsd": "tsd --version && tsd",
+    "check:dts": "./node_modules/typescript/bin/tsc --noEmit mongodb.d.ts && tsd",
+    "check:test": "mocha --file test/tools/runner test/integration",
+    "check:unit": "mocha test/unit/",
+    "check:ts": "./node_modules/typescript/bin/tsc -v && ./node_modules/typescript/bin/tsc --noEmit",
+    "check:atlas": "mocha --config \"test/manual/mocharc.json\" test/manual/atlas_connectivity.test.js",
+    "check:adl": "mocha --file test/tools/runner test/manual/data_lake.test.js",
+    "check:aws": "mocha --file test/tools/runner test/integration/auth/mongodb_aws.test.js",
+    "check:ocsp": "mocha --config \"test/manual/mocharc.json\" test/manual/ocsp_support.test.js",
+    "check:kerberos": "mocha --config \"test/manual/mocharc.json\" test/manual/kerberos.test.js",
+    "check:tls": "mocha --config \"test/manual/mocharc.json\" test/manual/tls_support.test.js",
+    "check:ldap": "mocha --config \"test/manual/mocharc.json\" test/manual/ldap.test.js",
+    "check:socks5": "mocha --config \"test/manual/mocharc.json\" test/manual/socks5.test.ts",
+    "check:csfle": "mocha --file test/tools/runner test/integration/client-side-encryption",
+    "check:snappy": "mocha test/unit/assorted/snappy.test.js",
+    "prepare": "node etc/prepare.js",
+    "release": "standard-version -i HISTORY.md",
+    "test": "npm run check:lint && npm run test:all",
+    "test:all": "npm run check:unit && npm run check:test"
+  },
+  "tsd": {
+    "directory": "test/types",
+    "compilerOptions": {
+      "strict": true,
+      "target": "esnext",
+      "module": "commonjs",
+      "moduleResolution": "node"
+    }
   }
 }
