@@ -19,66 +19,67 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "bson": "^1.1.4",
-    "kareem": "2.3.1",
-    "mongodb": "3.5.6",
-    "mongoose-legacy-pluralize": "1.0.2",
-    "mpath": "0.7.0",
-    "mquery": "3.2.2",
-    "ms": "2.1.2",
-    "regexp-clone": "1.0.0",
-    "safe-buffer": "5.1.2",
-    "sliced": "1.0.1",
-    "sift": "7.0.1"
+    "bson": "^4.2.2",
+    "kareem": "2.3.5",
+    "mongodb": "4.3.1",
+    "mpath": "0.8.4",
+    "mquery": "4.0.2",
+    "ms": "2.1.3",
+    "sift": "16.0.0"
   },
   "devDependencies": {
-    "acquit": "1.x",
-    "acquit-ignore": "0.1.x",
-    "acquit-require": "0.1.x",
-    "async": "2.6.2",
-    "babel-loader": "7.1.4",
-    "babel-preset-es2015": "6.24.1",
+    "@babel/core": "7.17.8",
+    "@babel/preset-env": "7.16.11",
+    "@typescript-eslint/eslint-plugin": "5.17.0",
+    "@typescript-eslint/parser": "5.17.0",
+    "acquit": "1.2.1",
+    "acquit-ignore": "0.2.0",
+    "acquit-require": "0.1.1",
+    "axios": "0.26.1",
+    "babel-loader": "8.2.4",
     "benchmark": "2.1.4",
-    "bluebird": "3.5.5",
-    "chalk": "2.4.2",
-    "cheerio": "1.0.0-rc.2",
-    "co": "4.6.0",
+    "bluebird": "3.7.2",
+    "cheerio": "1.0.0-rc.10",
     "dox": "0.3.1",
-    "eslint": "5.16.0",
-    "eslint-plugin-mocha-no-only": "1.1.0",
-    "highlight.js": "9.1.0",
+    "eslint": "8.12.0",
+    "eslint-plugin-mocha-no-only": "1.1.1",
+    "highlight.js": "11.5.0",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
-    "marked": "0.6.2",
-    "mocha": "5.x",
+    "marked": "4.0.12",
+    "mocha": "9.2.2",
     "moment": "2.x",
-    "mongodb-topology-manager": "1.0.11",
-    "mongoose-long": "0.2.1",
-    "node-static": "0.7.11",
-    "object-sizeof": "1.3.0",
-    "pug": "2.0.3",
+    "mongodb-memory-server": "^8.3.0",
+    "nyc": "^15.1.0",
+    "pug": "3.0.2",
     "q": "1.5.1",
-    "semver": "5.5.0",
-    "uuid": "2.0.3",
-    "uuid-parse": "1.0.0",
-    "validator": "10.8.0",
-    "webpack": "4.16.4"
+    "serve-handler": "6.1.3",
+    "tsd": "0.20.0",
+    "typescript": "4.6.3",
+    "uuid": "8.3.2",
+    "webpack": "4.44.1"
   },
   "directories": {
     "lib": "./lib/mongoose"
   },
   "scripts": {
     "lint": "eslint .",
+    "lint-js": "eslint . --ext .js",
+    "lint-ts": "eslint . --ext .ts",
     "build-browser": "node build-browser.js",
     "prepublishOnly": "npm run build-browser",
     "release": "git pull && git push origin master --tags && npm publish",
-    "release-legacy": "git pull origin 4.x && git push origin 4.x --tags && npm publish --tag legacy",
-    "test": "mocha --exit",
-    "test-cov": "nyc --reporter=html --reporter=text npm test"
+    "release-legacy": "git pull origin 5.x && git push origin 5.x --tags && npm publish --tag legacy",
+    "mongo": "node ./tools/repl.js",
+    "test": "mocha --exit ./test/*.test.js",
+    "test-tsd": "node ./test/types/check-types-filename && tsd",
+    "tdd": "mocha ./test/*.test.js ./test/typescript/main.test.js --inspect --watch --recursive --watch-files ./**/*.js",
+    "test-coverage": "nyc --reporter=html --reporter=text npm test"
   },
   "main": "./index.js",
+  "types": "./types/index.d.ts",
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=12.0.0"
   },
   "bugs": {
     "url": "https://github.com/Automattic/mongoose/issues/new"
