@@ -1,6 +1,6 @@
 {
   "name": "mquery",
-  "version": "3.2.2",
+  "version": "4.0.2",
   "description": "Expressive query building for MongoDB",
   "main": "lib/mquery.js",
   "scripts": {
@@ -13,19 +13,16 @@
     "url": "git://github.com/aheckmann/mquery.git"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=12.0.0"
   },
   "dependencies": {
-    "bluebird": "3.5.1",
-    "debug": "3.1.0",
-    "regexp-clone": "^1.0.0",
-    "safe-buffer": "5.1.2",
-    "sliced": "1.0.1"
+    "debug": "4.x"
   },
   "devDependencies": {
-    "eslint": "5.x",
-    "mocha": "4.1.0",
-    "mongodb": "3.1.1"
+    "eslint": "8.x",
+    "eslint-plugin-mocha-no-only": "1.1.1",
+    "mocha": "9.x",
+    "mongodb": "4.x"
   },
   "bugs": {
     "url": "https://github.com/aheckmann/mquery/issues/new"
@@ -37,48 +34,5 @@
     "query",
     "builder"
   ],
-  "homepage": "https://github.com/aheckmann/mquery/",
-  "eslintConfig": {
-    "env": {
-      "node": true,
-      "mocha": true,
-      "es6": false
-    },
-    "extends": "eslint:recommended",
-    "parserOptions": {
-      "ecmaVersion": 5
-    },
-    "rules": {
-      "comma-style": "error",
-      "consistent-this": [
-        "error",
-        "_this"
-      ],
-      "indent": [
-        "error",
-        2,
-        {
-          "SwitchCase": 1,
-          "VariableDeclarator": 2
-        }
-      ],
-      "keyword-spacing": "error",
-      "no-console": "off",
-      "no-multi-spaces": "error",
-      "func-call-spacing": "error",
-      "no-trailing-spaces": "error",
-      "quotes": [
-        "error",
-        "single"
-      ],
-      "semi": "error",
-      "space-before-blocks": "error",
-      "space-before-function-paren": [
-        "error",
-        "never"
-      ],
-      "space-infix-ops": "error",
-      "space-unary-ops": "error"
-    }
-  }
+  "homepage": "https://github.com/aheckmann/mquery/"
 }
