@@ -1,11 +1,40 @@
 import React, { useState, useEffect } from 'react';
 import { Jumbotron, Container, Col, Form, Button, Card, CardColumns } from 'react-bootstrap';
-
-import Auth from '../utils/auth';
-import { saveBook, searchGoogleBooks } from '../utils/API';
+import { useQuery, useMutation } from '@apollo/client';
+import { ADD_BOOK } from '../utils/mutations';
+import { QUERY_ME, QUERY_BOOKS } from '../utils/queries';
+import { searchGoogleBooks } from '../utils/API';
 import { saveBookIds, getSavedBookIds } from '../utils/localStorage';
+import auth from '../utils/auth';
 
 const SearchBooks = () => {
+  // const [addBook, { error }] = useMutation(ADD_BOOK);
+  const [addBook, { error }] = useMutation(ADD_BOOK, {
+    update(cache, { data: { addBook } }) {
+      try {
+        // could potentially not exist yet, so wrap in a try...catch
+        const { books } = cache.readQuery({ query: QUERY_BOOKS });
+        cache.writeQuery({
+          query: QUERY_BOOKS,
+          data: { books: [addBook, ...books] }
+        });
+      } catch (e) {
+        console.error(e);
+      }
+  
+      // update me object's cache, appending new thought to the end of the array
+      const { me } = cache.readQuery({ query: QUERY_ME });
+      cache.writeQuery({
+        query: QUERY_ME,
+        data: { me: { ...me, savedBooks: [...me.savedBooks, addBook] } }
+      });
+    }
+  });
+
+   // use object destructuring to extract `data` from the `useQuery` Hook's response and rename it `userData` to be more descriptive
+   const { data: userData } = useQuery(QUERY_ME);
+  //  console.log(userData)
+
   // create state for holding returned google api data
   const [searchedBooks, setSearchedBooks] = useState([]);
   // create state for holding our search field data
@@ -19,7 +48,7 @@ const SearchBooks = () => {
   useEffect(() => {
     return () => saveBookIds(savedBookIds);
   });
-
+  
   // create method to search for books and set state on form submit
   const handleFormSubmit = async (event) => {
     event.preventDefault();
@@ -40,9 +69,10 @@ const SearchBooks = () => {
       const bookData = items.map((book) => ({
         bookId: book.id,
         authors: book.volumeInfo.authors || ['No author to display'],
-        title: book.volumeInfo.title,
-        description: book.volumeInfo.description,
+        title: book.volumeInfo.title || '',
+        description: book.volumeInfo.description || '',
         image: book.volumeInfo.imageLinks?.thumbnail || '',
+        link: book.volumeInfo.infoLink || ''
       }));
 
       setSearchedBooks(bookData);
@@ -54,27 +84,57 @@ const SearchBooks = () => {
 
   // create function to handle saving a book to our database
   const handleSaveBook = async (bookId) => {
+   
     // find the book in `searchedBooks` state by the matching id
     const bookToSave = searchedBooks.find((book) => book.bookId === bookId);
-
+    console.log(bookToSave);
+    if (bookToSave.bookId) {
+      var newBookId = bookToSave.bookId;
+    } else {
+      newBookId = "";
+    }
+    if (bookToSave.authors) {
+      var newAuthors = bookToSave.authors;
+    } else {
+      newAuthors = "";
+    }
+    if (bookToSave.title) {
+      var newTitle = bookToSave.title;
+    } else {
+      newTitle = "";
+    }
+    if (bookToSave.description) {
+      var newDescription = bookToSave.description;
+    } else {
+      newDescription = "";
+    }
+    if (bookToSave.link) {
+      var newLink = bookToSave.link;
+    } else {
+      newLink = "";
+    }
+    if (bookToSave.image) {
+      var newImage = bookToSave.image;
+    } else {
+      newImage = "";
+    }
+  
     // get token
-    const token = Auth.loggedIn() ? Auth.getToken() : null;
+    const token = auth.loggedIn() ? auth.getToken() : null;
 
     if (!token) {
       return false;
     }
-
+  
     try {
-      const response = await saveBook(bookToSave, token);
+      // add thought to database
+      await addBook({
+        variables: { bookId: newBookId, authors: newAuthors, title: newTitle, description: newDescription, link: newLink, image: newImage },
+      });
 
-      if (!response.ok) {
-        throw new Error('something went wrong!');
-      }
-
-      // if book successfully saves to user's account, save book id to state
       setSavedBookIds([...savedBookIds, bookToSave.bookId]);
-    } catch (err) {
-      console.error(err);
+    } catch (e) {
+        console.error(e);
     }
   };
 
@@ -122,7 +182,13 @@ const SearchBooks = () => {
                   <Card.Title>{book.title}</Card.Title>
                   <p className='small'>Authors: {book.authors}</p>
                   <Card.Text>{book.description}</Card.Text>
-                  {Auth.loggedIn() && (
+                  <Button 
+                    className='btn-block btn-info'
+                    onClick={() => window.location.href = `${book.link}`}
+                  >
+                    Link
+                  </Button>
+                  {auth.loggedIn() && (
                     <Button
                       disabled={savedBookIds?.some((savedBookId) => savedBookId === book.bookId)}
                       className='btn-block btn-info'
