@@ -2,11 +2,11 @@ const express = require('express');
 const { ApolloServer } = require('apollo-server-express');
 const { authMiddleware } = require('./utils/auth');
 const path = require('path');
-const db = require('./config/connection');
 // const routes = require('./routes');
 
 // import our typeDefs and resolvers
 const { typeDefs, resolvers } = require('./schemas');
+const db = require('./config/connection');
 
 const app = express();
 const PORT = process.env.PORT || 3001;
@@ -41,9 +41,9 @@ if (process.env.NODE_ENV === 'production') {
 }
 
 // app.use(routes);
-app.get('*', (req, res) => {
-  res.sendFile(path.join(__dirname, '../client/build/index.html'));
-});
+// app.get('*', (req, res) => {
+//   res.sendFile(path.join(__dirname, '../client/build/index.html'));
+// });
 
 db.once('open', () => {
   app.listen(PORT, () => console.log(`🌍 Now listening on localhost:${PORT}`));
